Added some 'base' docker files for clusternator projects.
The idea is that if you're using the node/express stack you don't want to do all the apt-get stuff everytime you build. Same thing for downloading and installing node.

So there's a base 'ubuntu + apt-get + set up swuser' image
Then there are node 4 and 5 images that depend on the base one
And finally your clusternated project should depend on one of the node ones.

Then all that needs to be done in CI is npm install, serve.sh.